+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/search/builder/models/PredicateBuilderTypeEnum.java
@@ -1,0 +1,5 @@
+package ca.uhn.fhir.jpa.search.builder.models;
+
+public enum PredicateBuilderTypeEnum {
+	DATE, COORDS, NUMBER, QUANTITY, REFERENCE, SOURCE, STRING, TOKEN, TAG
+}
